lab2/requirements.txt
Pillow==7.0.0
tabulate==0.8.6